+++ frontend/app/src/main/java/com/example/frontend/MissionActivity.kt
@@ -28,6 +28,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -46,6 +47,7 @@ import androidx.compose.ui.window.Dialog
 import com.example.frontend.api.MissionService
 import com.example.frontend.model.MissionModel
 import com.example.frontend.repository.UserContextRepository
+import com.example.frontend.repository.defaultMissions
 import com.example.frontend.ui.theme.FrontendTheme
 import com.example.frontend.usecase.MissionUseCase
 import retrofit2.Call
@@ -53,26 +55,26 @@ import retrofit2.Callback
 import retrofit2.Response
 
 class MissionActivity : ComponentActivity() {
-    val userContextRepository = UserContextRepository(this)
     var missions by mutableStateOf<List<MissionModel>>(emptyList())
     @OptIn(ExperimentalMaterial3Api::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            missions = MissionUseCase(this).fetch()
-            FrontendTheme {
+            val missionUseCase = MissionUseCase(this)
 
+            LaunchedEffect(Unit) {
+                missionUseCase.fetch { fetchedMissions ->
+                    missions = fetchedMissions
+                }
+            }
+            FrontendTheme {
                 ShowMissionUI(missions = missions) {
                     // Handle switch to register
                 }
             }
         }
     }
-
-
 }
-
-
 
 @Composable
 fun ShowMissionUI(missions: List<MissionModel>, onSwitchToRegister: () -> Unit) {
@@ -100,7 +102,6 @@ fun ShowMissionUI(missions: List<MissionModel>, onSwitchToRegister: () -> Unit) 
                     .clickable {
                         val nextIntent = Intent(context, MapActivity::class.java)
                         context.startActivity(nextIntent)
-                        // finish current activity
                         if (context is Activity) {
                             context.finish()
                         }
@@ -119,7 +120,6 @@ fun ShowMissionUI(missions: List<MissionModel>, onSwitchToRegister: () -> Unit) 
                 )
             )
         }
-
         if (missions.isEmpty()) {
             Spacer(modifier = Modifier.height(16.dp))
             CircularProgressIndicator(
@@ -298,16 +298,6 @@ fun ShowMoreDescriptionDialog(
     }
 }
 
-val defaultMissions = listOf(
-    MissionModel("미션1", "친구와 우연히 만나기", false, "예상치 못한 장소에서 친구와 마주쳐 보세요!"),
-    MissionModel("미션2", "친구와 약속 잡기", false, "친구와 약속을 잡아 보세요!"),
-    MissionModel("미션3", "친구와 약속 장소 정하기", false, "3명 이상의 친구와 약속을 잡아 보세요!"),
-    MissionModel("미션4", "자하연 근처에서 친구 마주치기", false, "자하연에서 친구와 마주쳐 보세요!"),
-    MissionModel("미션5", "관악산 등산하기", false, "관악산에 올라가 보세요!"),
-    MissionModel("미션6", "도서관에 한 시간 머물기", false, "도서관에 머물며 책을 읽는 시간을 가져 보세요!"),
-    MissionModel("미션7", "친구 세 명과 만나기", false, "친구 세 명과 약속을 잡아 보세요!"),
-    MissionModel("미션8", "친구 스무 명 추가하기", false, "친구 20 명을 추가해 보세요!")
-)
 
 @Preview(showBackground = true)
 @Composable
