+++ frontend/app/src/main/java/com/example/frontend/usecase/MissionUseCase.kt
@@ -1,13 +1,15 @@
 package com.example.frontend.usecase
 
 import android.content.Context
+import android.widget.Toast
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import com.example.frontend.api.AuthService
 import com.example.frontend.api.MissionService
-import com.example.frontend.defaultMissions
+
 import com.example.frontend.model.MissionModel
+import com.example.frontend.repository.defaultMissions
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -16,8 +18,7 @@ class MissionUseCase(
     private val context: Context,
     private val missionService: MissionService = MissionService.create()
 ) {
-        fun fetch(): List<MissionModel> {
-        var missions by mutableStateOf<List<MissionModel>>(emptyList())
+    fun fetch(onMissionsFetched: (List<MissionModel>) -> Unit) {
         val call = missionService.getMissionList()
 
         call.enqueue(object : Callback<List<MissionModel>> {
@@ -26,17 +27,16 @@ class MissionUseCase(
                 response: Response<List<MissionModel>>
             ) {
                 if (response.isSuccessful) {
-                    missions = response.body() ?: emptyList()
-
+                    val missions = response.body() ?: emptyList()
+                    onMissionsFetched(missions)
                 } else {
-
+                    onMissionsFetched(defaultMissions)
                 }
             }
 
             override fun onFailure(call: Call<List<MissionModel>>, t: Throwable) {
-                    missions = defaultMissions
+                onMissionsFetched(defaultMissions)
             }
         })
-            return missions
     }
 }
